feat: Add CONTAINERISE env for Docker setup
Introduce a new environment variable `CONTAINERISE` to determine if Singularity should be used in a Docker environment. Update the `nextflow_command_string` in `api_utils.py` to check the value of `CONTAINERISE`.

This is necessary to allow flexibility when running the application in a Docker environment that may not require or support Singularity.